Change logic of playRound and playGame functions
Before this change, the playGame function was needed to play
the game; you couldn't play one round by calling the playRound
function. It simply did not work. Change the logic around so all things
happen inside the playRound function, and the playGame func just
repeats the playRound func and gives the end result. Still, the playGame
func for now, is commented out.